+++ cypress/support/commands.ts
@@ -1,2 +1,2 @@
 import '@ui5/webcomponents-cypress-commands';
-import "../../i18n";
+import "../../src/utils/i18n/i18n";
