+++ src/components/Dialogs/CreateProjectDialogContainer.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useRef } from 'react';
+import { useCallback, useEffect, useMemo, useRef } from 'react';
 import { useApiResourceMutation } from '../../lib/api/useApiResource';
 import { ErrorDialogHandle } from '../Shared/ErrorMessageBox.tsx';
 import { APIError } from '../../lib/api/error';
@@ -12,7 +12,7 @@ import { useTranslation } from 'react-i18next';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 import { CreateProject, CreateProjectResource, CreateProjectType } from '../../lib/api/types/crate/createProject.ts';
-import { validationSchemaProjectWorkspace } from '../../lib/api/validations/schemas.ts';
+import { createProjectWorkspaceSchema } from '../../lib/api/validations/schemas.ts';
 import { CreateDialogProps } from './CreateWorkspaceDialogContainer.tsx';
 
 export function CreateProjectDialogContainer({
@@ -22,6 +22,8 @@ export function CreateProjectDialogContainer({
   isOpen: boolean;
   setIsOpen: (isOpen: boolean) => void;
 }) {
+  const { t } = useTranslation();
+  const validationSchemaProjectWorkspace = useMemo(() => createProjectWorkspaceSchema(t), [t]);
   const {
     watch,
     register,
@@ -39,7 +41,6 @@ export function CreateProjectDialogContainer({
       members: [],
     },
   });
-  const { t } = useTranslation();
   const { user } = useAuthOnboarding();
 
   const username = user?.email;
